+++ tests/test_orm_model.py
@@ -68,6 +68,47 @@ def test_model_empty_meta_with_callable():
         m.assert_not_called()
 
 
+def test_model_meta_dict():
+    """
+    Test that we can define Meta as a dict rather than a class.
+    """
+
+    class Address(Model):
+        Meta = {
+            "api_key": "fake_api_key",
+            "base_id": "fake_base_id",
+            "table_name": "fake_table_name",
+            "timeout": (1, 1),
+            "retry": False,
+        }
+
+    assert Address.meta.api.api_key == "fake_api_key"
+
+
+@pytest.mark.parametrize("invalid_meta", ([1, 2, 3], "invalid", True))
+def test_model_invalid_meta(invalid_meta):
+    """
+    Test that model creation raises a TypeError if Meta is an invalid type.
+    """
+    with pytest.raises(TypeError):
+
+        class Address(Model):
+            Meta = invalid_meta
+
+
+@pytest.mark.parametrize("meta_kwargs", [{"timeout": 1}, {"retry": "sure"}])
+def test_model_meta_checks_types(meta_kwargs):
+    """
+    Test that accessing meta raises a TypeError if a value is an invalid type.
+    """
+
+    class Address(Model):
+        Meta = fake_meta(**meta_kwargs)
+
+    with pytest.raises(TypeError):
+        Address.meta.api
+
+
 @pytest.mark.parametrize("name", ("exists", "id"))
 def test_model_overlapping(name):
     """
@@ -178,8 +219,8 @@ def test_from_ids(mock_api):
     contacts = FakeModel.from_ids(fake_ids)
     mock_api.assert_called_once_with(
         method="get",
-        url=FakeModel.get_table().url,
-        fallback=("post", FakeModel.get_table().url + "/listRecords"),
+        url=FakeModel.meta.table.url,
+        fallback=("post", FakeModel.meta.table.url + "/listRecords"),
         options={
             "formula": "OR(%s)" % ", ".join(f"RECORD_ID()='{id}'" for id in fake_ids)
         },
@@ -248,7 +289,7 @@ def test_get_fields_by_id(fake_records_by_id):
     """
     with Mocker() as mock:
         mock.get(
-            f"{FakeModelByIds.get_table().url}?&returnFieldsByFieldId=1&cellFormat=json",
+            f"{FakeModelByIds.meta.table.url}?&returnFieldsByFieldId=1&cellFormat=json",
             json=fake_records_by_id,
             complete_qs=True,
             status_code=200,
@@ -286,7 +327,7 @@ def test_dynamic_model_meta():
     f = Fake()
     Fake.Meta.table_name.assert_not_called()
 
-    assert f._get_meta("api_key") == data["api_key"]
-    assert f._get_meta("base_id") == data["base_id"]
-    assert f._get_meta("table_name") == data["table_name"]
+    assert f.meta.get("api_key") == data["api_key"]
+    assert f.meta.get("base_id") == data["base_id"]
+    assert f.meta.get("table_name") == data["table_name"]
     Fake.Meta.table_name.assert_called_once()