+++ docs/source/changelog.rst
@@ -26,6 +26,7 @@ Changelog
   - `PR #362 <https://github.com/gtalarico/pyairtable/pull/362>`_.
 * Added ORM fields that :ref:`require a non-null value <Required Values>`.
   - `PR #363 <https://github.com/gtalarico/pyairtable/pull/363>`_.
+* Refactored methods for accessing ORM model configuration.
 
 2.3.3 (2024-03-22)
 ------------------------